lib/requirements.txt
PyYAML==6.0.1
jsonschema==4.23.0
requests==2.32.3
